services/gateway/requirements.txt
fastapi
uvicorn
google-cloud-firestore
google-cloud-pubsub
google-api-core
pydantic>=2email-validator
